+++ segmentation/include/pcl/segmentation/impl/extract_polygonal_prism_data.hpp
@@ -157,7 +157,9 @@ pcl::ExtractPolygonalPrismData<PointT>::segment (pcl::PointIndices &output)
 
   if (static_cast<int> (planar_hull_->size ()) < min_pts_hull_)
   {
-    PCL_ERROR ("[pcl::%s::segment] Not enough points (%lu) in the hull!\n", getClassName ().c_str (), planar_hull_->size ());
+    PCL_ERROR("[pcl::%s::segment] Not enough points (%zu) in the hull!\n",
+              getClassName().c_str(),
+              static_cast<std::size_t>(planar_hull_->size()));
     output.indices.clear ();
     return;
   }
